자취왕 채팅 오류 수정 이메일 로그아웃 오류 수정 휴대폰 인증 타이머 에러 수정 댓글 삭제 기능 추가
*해야할 일*
프로필 사진 변경 구현(아직 완벽하지 않음)
게시글 수정, 삭제 구현
게시판, 채팅 아이템 레이아웃 수정
게시판 사진 올리기
프로필쪽 내프로필, 로그아웃만 구현했음
알림기능 구현해야함

+++ app/src/main/java/com/example/capstondesign/model/Comment_Adapter.java
@@ -4,6 +4,8 @@ import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,12 +16,17 @@ import android.widget.Toast;
 
 import com.example.capstondesign.R;
 import com.example.capstondesign.controller.Fragment_board;
+import com.example.capstondesign.controller.FreeBoard;
+import com.example.capstondesign.controller.LoginAcitivity;
 
 import java.util.ArrayList;
+import java.util.concurrent.ExecutionException;
 
 public class Comment_Adapter extends BaseAdapter implements View.OnClickListener {
     private Context mContext;
     private Activity mActivity;
+    Profile profile = LoginAcitivity.profile;
+    String nickname;
     private ArrayList<Comment_Item> arr;
     private int pos;
     private Fragment_board ma;
@@ -29,7 +36,6 @@ public class Comment_Adapter extends BaseAdapter implements View.OnClickListener
         this.mActivity = mActivity;
         this.arr = arr_item;
         this.ma = mc;
-//		myFont = Typeface.createFromAsset(mContext.getAssets(), "BareunDotum.ttf");
     }
     @Override
     public int getCount() {
@@ -52,6 +58,7 @@ public class Comment_Adapter extends BaseAdapter implements View.OnClickListener
         }
         pos = position;
         if(arr.size() != 0){
+            getNick();
             TextView nickname_text = (TextView)convertView.findViewById(R.id.nickname_text);
             nickname_text.setText(arr.get(pos).getNick());
             TextView content_text = (TextView)convertView.findViewById(R.id.content_text);
@@ -64,15 +71,32 @@ public class Comment_Adapter extends BaseAdapter implements View.OnClickListener
         return convertView;
     }
     public void onClick(View v){
+        DeleteCommentTask deleteCommentTask = new DeleteCommentTask();
         final int tag = Integer.parseInt(v.getTag().toString());
+        Log.d("FFFFFFFF", nickname);
         switch(v.getId()){
             case R.id.delete_btn:
                 AlertDialog.Builder alertDlg = new AlertDialog.Builder(mActivity);
                 alertDlg.setPositiveButton("예", new DialogInterface.OnClickListener(){
                     @Override
                     public void onClick( DialogInterface dialog, int which ) {
-                        deleteArr(tag);
-                        Toast.makeText(mContext, "삭제되었습니다.", Toast.LENGTH_LONG).show();
+
+                        String result = null;
+                        try {
+                            result = deleteCommentTask.execute(arr.get(pos).getNick(), FreeBoard.title , arr.get(pos).getComment()).get();
+                            if(arr.get(pos).getNick().equals(nickname)) {
+                                if (result.contains("delete")) {
+                                    deleteArr(tag);
+                                    Toast.makeText(mContext, "삭제되었습니다.", Toast.LENGTH_SHORT).show();
+                                }
+                            } else {
+                                Toast.makeText(mContext, "본인의 댓글만 삭제할 수 있습니다.", Toast.LENGTH_SHORT).show();
+                            }
+                        } catch (ExecutionException e) {
+                            e.printStackTrace();
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
                     }
                 });
                 alertDlg.setNegativeButton("아니오", null);
@@ -87,4 +111,18 @@ public class Comment_Adapter extends BaseAdapter implements View.OnClickListener
         arr.remove(p);
         notifyDataSetChanged();
     }
+
+    void getNick() {
+        ProfileTask profileTask = new ProfileTask();
+        try {
+            String result = profileTask.execute(profile.getName(), profile.getEmail()).get();
+            nickname = profileTask.substringBetween(result, "nickname:", "/");
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } catch (ExecutionException e) {
+            e.printStackTrace();
+        }
+    }
+
+
 }
